+++ src/index.ts
@@ -3,8 +3,10 @@
  */
 export * from "./error/coreError";
 export * from "./error/networkError";
+export * from "./error/platformError";
 export * from "./factories/factoryBase";
 export * from "./factories/networkClientFactory";
+export * from "./factories/platformCryptoFactory";
 export * from "./factories/rngServiceFactory";
 export * from "./helpers/arrayHelper";
 export * from "./helpers/errorHelper";
@@ -16,9 +18,10 @@ export * from "./interfaces/IBackgroundTaskService";
 export * from "./interfaces/ILogger";
 export * from "./interfaces/INetworkClient";
 export * from "./interfaces/INetworkEndPoint";
-export * from "./interfaces/networkProtocol";
+export * from "./interfaces/IPlatformCrypto";
 export * from "./interfaces/IRngService";
 export * from "./interfaces/ITimeService";
+export * from "./interfaces/networkProtocol";
 export * from "./loggers/consoleLogger";
 export * from "./loggers/nullLogger";
 export * from "./network/networkEndPoint";